Audit hook dialog UX copy
- Replace env-var-heavy placeholders with plain-language agent prompts
  for start and review hooks (start uses the complete-and-review prompt)
- Add a dialog-level CLI hint (info icon + tooltip) explaining hooks run
  in the task worktree and agents can use the ouijit CLI
- Add an explainer blurb to the environment variables disclosure
- Extract shared HookCliHint and HookEnvVars components, removing the
  triplicated env-var list and copy logic across the three hook dialogs
- Raise tooltip portal z-index above modal dialogs so hints aren't hidden
- Align the onboarding example hook to the canonical start prompt
