+++ src/components/ui/Tooltip.tsx
@@ -73,7 +73,7 @@ export function Tooltip({
         createPortal(
           <div
             ref={refs.setFloating}
-            className="fixed z-[10000] pointer-events-none"
+            className="fixed z-[10002] pointer-events-none"
             style={floatingStyles}
             {...getFloatingProps()}
           >
