+++ src/components/ui/TooltipButton.tsx
@@ -68,7 +68,7 @@ export function TooltipButton({
         createPortal(
           <div
             ref={refs.setFloating}
-            className="fixed z-[10000] pointer-events-none"
+            className="fixed z-[10002] pointer-events-none"
             style={floatingStyles}
             {...getFloatingProps()}
           >
